restructured the project into one python package, group relevent python modules into small packages
