+++ raycaster/__main__.py
@@ -1,4 +1,4 @@
-from game import Game
+from .game.game import Game
 
 
 if __name__ == "__main__":
