+++ raycaster/assets/objects/static/candlebra.png
