+++ raycaster/assets/objects/static/npc.png
