+++ raycaster/assets/walls/3.png
