+++ raycaster/game/__init__.py
@@ -1,0 +1,3 @@
+from .asset_loader import AssetLoader
+from .map import Map
+from .player import Player
