+++ raycaster/objects/__init__.py
@@ -1,0 +1,2 @@
+from .object_manager import ObjectManager
+from .sprite_object import SpriteObject
