+++ raycaster/objects/object_manager.py
@@ -1,11 +1,11 @@
 from typing import TYPE_CHECKING
 
-from settings import Settings
-from asset_loader import AssetLoader
-from sprite_object import SpriteObject
+from ..core import Settings
+from ..game import AssetLoader
+from .sprite_object import SpriteObject
 
 if TYPE_CHECKING:
-    from player import Player
+    from ..game import Player
 
 
 class ObjectManager:
