+++ raycaster/objects/sprite_object.py
@@ -2,12 +2,11 @@ from typing import TYPE_CHECKING
 import math
 import pygame
 
-from updatable import Updatable
-from settings import Settings
-from utils import calculate_distance
+from ..core import Updatable, Settings
+from ..utils import calculate_distance
 
 if TYPE_CHECKING:
-    from player import Player
+    from ..game import Player
 
 
 class SpriteObject(Updatable):
