+++ raycaster/rendering/ray.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from settings import Settings
+from ..core import Settings
 
 
 @dataclass
